+++ puffin-imgui/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+* Add slider for controlling number of frames recorded.
+* Show total frames recorded and their total size.
 
 
 ## 0.13.4 - 2021-11-05