+++ puffin_egui/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the egui crate will be documented in this file.
 
 
 ## Unreleased
+* Show total frames recorded and their total size.
 
 
 ## 0.10.3 - 2021-11-08